server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.21.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pycryptodome==3.21.0
httpx==0.27.2
python-multipart==0.0.12
